requirements.txt
affine==2.3.0
appdirs==1.4.4
appnope==0.1.2
astropy==6.0.0
astropy-iers-data==0.2024.1.8.0.30.55
attrs==23.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bs4==0.0.1
Cartopy @ git+https://github.com/SciTools/cartopy.git@64ac1728b89b01ed41b609d636a6b7f998a70958
celest==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.7
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
contextily==1.2.0
cycler==0.10.0
Cython==0.29.25
debugpy==1.4.3
decorator==5.1.0
distlib==0.3.2
docker-pycreds==0.4.0
entrypoints==0.3
exceptiongroup==1.2.0
fastjsonschema==2.19.1
filelock==3.0.12
findspark==2.0.1
Fiona==1.8.20
Flask==2.0.1
Flask-Login==0.5.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
fsspec==2023.10.0
geographiclib==1.52
geopandas==0.9.0
geopy==2.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==1.1.1
huggingface-hub==0.19.4
idna==3.3
iniconfig==2.0.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jplephem==2.19
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julian==0.14
jupyter-client==7.0.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.3.2
# Editable install with no version control (mapnik==3.0.23)
-e /Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages/mapnik-3.0.23-py3.9-macosx-10.9-x86_64.egg
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mercantile==1.2.1
mpmath==1.3.0
munch==2.5.1.dev12
mysql-connector-python==8.0.27
nbformat==5.10.4
nest-asyncio==1.5.1
networkx==3.1
numpy==1.26.3
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==1.3.4
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==4.2.1
plotly==5.18.0
pluggy==1.4.0
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.9.8
ptyprocess==0.7.0
py4j==0.10.9.3
pyerfa==2.0.1.1
pygame==2.5.2
Pygments==2.10.0
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==2.4.7
pyproj==3.3.0
pyshp==2.1.3
pyspark==3.2.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2021.3
PyYAML==6.0.1
pyzmq==22.3.0
rasterio==1.2.10
referencing==0.35.0
regex==2023.10.3
requests==2.26.0
rpds-py==0.18.0
safetensors==0.4.0
scikit-learn==1.0.1
scipy==1.7.3
SCons==4.2.0
seaborn==0.11.2
sentry-sdk==2.1.1
setproctitle==1.3.3
Shapely==1.8.0
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
soupsieve==2.2.1
SQLAlchemy==1.4.25
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.0.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1
tornado==6.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.2
typing_extensions==4.8.0
urllib3==1.26.18
virtualenv==20.7.2
virtualenv-clone==0.5.6
wandb==0.16.6
wcwidth==0.2.5
Werkzeug==2.0.1
xyzservices==2021.11.0
